+++ src/BigNumber.php
@@ -417,6 +417,7 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
      *
      * If the number is greater than the largest representable floating point number, positive infinity is returned.
      * If the number is less than the smallest representable floating point number, negative infinity is returned.
+     * This method never returns NaN.
      *
      * @pure
      */